requirements.txt
## The following requirements were added by pip freeze:
aiofiles==0.4.0
aiogoogletrans==3.2.1
aiohttp==3.4.4
aioredis==1.2.0
APScheduler==3.5.3
arsenic==1.0.0.dev8
async-timeout==3.0.0
asyncpg==0.18.2
attrs==18.2.0
blinker==1.3
cachetools==2.1.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloud-init==0.7.9
configobj==5.0.6
cryptography==1.7.1
dblpy==0.1.6
discord.py==1.0.0a1590+g860d6a9
entrypoints==0.2.3.post3
Flask==1.0.2
Flask-Limiter==1.0.1
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
hiredis==0.2.0
httplib2==0.11.3
idna==2.7
idna-ssl==1.1.0
itsdangerous==0.24
Jinja2==2.10
jishaku==0.1.2
jsonify==0.5
jsonpatch==1.10
jsonpointer==1.10
keyring==15.1.0
keyrings.alt==3.1
limits==1.3
MarkupSafe==1.0
multidict==4.4.2
oauthlib==2.0.1
Pillow==5.3.0
prettytable==0.7.2
psutil==5.4.7
pyasn1==0.1.9
pyasn1-modules==0.2.2
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.1
PyGObject==3.30.1
PyJWT==1.4.2
PyNaCl==1.2.1
python-apt==1.6.2
python-resize-image==1.1.18
pytz==2018.6
pyxdg==0.25
PyYAML==3.12
ratelimiter==1.2.0.post0
requests==2.20.0
rsa==4.0
SecretStorage==2.3.1
six==1.10.0
speedtest-cli==2.0.2
structlog==18.2.0
tzlocal==1.5.1
ufw==0.35
unattended-upgrades==0.1
uritemplate==3.0.0
urllib3==1.24
uvloop==0.11.2
uWSGI==2.0.17.1
websockets==6.0
Werkzeug==0.14.1
yandex-translater==1.0
yarl==1.2.6
